+++ renderer/src/pages/Node/components/NodeVersion/index.module.scss
@@ -1,12 +1,3 @@
-.container {
-  padding: 16px 30px 0;
-
-  .cannelBtn {
-    margin-bottom: 20px;
-    float: right;
-  }
-}
-
 .row {
   margin-bottom: 32px;
   display: flex;
